Create a stateless CategoryFilter.js
